// apps/sync-service/src/openapi/transaction.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * SMILE
 * OpenAPI spec version: 1.0.0
 */
import axios from 'axios';
import type {
  AxiosRequestConfig,
  AxiosResponse
} from 'axios';

export type PostV2TransactionsBodyItemBatch = {
  code: string;
  expired_date: string;
  production_date: string;
  manufacture_id?: number;
};

export type PostV2TransactionsBodyItemTransactionPatientsItemOtherSequencesItem = {
  sequence: number;
  date: string;
  title?: string;
  method: number;
};

export type PostV2TransactionsBodyItemTransactionPatientsItem = {
  /** @nullable */
  vaccine_sequence: number | null;
  identity_type: number;
  patient_id: string;
  phone_number?: string;
  other_sequences?: PostV2TransactionsBodyItemTransactionPatientsItemOtherSequencesItem[];
};

export type PostV2TransactionsBodyItem = {
  activity_id: number;
  transaction_id?: number[];
  transaction_type_id: number;
  transaction_reason_id?: number;
  other_reason?: string;
  status_id?: number;
  entity_id: number;
  material_id: number;
  stock_id: number;
  customer_id: number;
  change_qty: number;
  broken_qty?: number;
  vaccine_sequence?: number;
  actual_transaction_date?: string;
  dose_1?: number;
  dose_2?: number;
  booster?: number;
  is_batches?: boolean;
  batch?: PostV2TransactionsBodyItemBatch;
  open_vial?: number;
  close_vial?: number;
  broken_open_vial?: number;
  broken_close_vial?: number;
  transaction_patients?: PostV2TransactionsBodyItemTransactionPatientsItem[];
  /** @nullable */
  preexposure_method?: number | null;
  year?: number;
  price?: number;
  source_material_id?: number;
};

export type PostV2Transactions200DataItem = {
  id: number;
  purchase_id?: number;
  batch_id?: number;
  stock_id?: number;
  stock_extermination_id?: number;
  injection_id?: number;
  discard_id?: number;
};

export type PostV2Transactions200 = {
  message: string;
  data: PostV2Transactions200DataItem[];
};

export const getSmile = () => {
/**
 * @summary v2/transactions
 */
const postV2Transactions = <TData = AxiosResponse<PostV2Transactions200>>(
    postV2TransactionsBodyItem: PostV2TransactionsBodyItem[], options?: AxiosRequestConfig
 ): Promise<TData> => {
    return axios.post(
      `/v2/transactions`,
      postV2TransactionsBodyItem,options
    );
  }

return {postV2Transactions}};
export type PostV2TransactionsResult = AxiosResponse<PostV2Transactions200>